add some validation

+++ controllers/car.js
@@ -1,5 +1,6 @@
 let db = require('../db/car');
 let helper = require('../helpers/car');
+let validation = require('../helpers/validation');
 
 exports.getAllCars = function(req, res) {
     let cars = db.getAll();
@@ -13,24 +14,41 @@ exports.createCar = function(req, res) {
 };
 
 exports.getCarById = function(req, res) {
-    let car = db.getById(req.params.carId);
-    showPromise(req, res, car);
+    let carId = parseInt(req.params.carId, 10);
+    let isValidCarId = validation.validateCarId(carId);
+    if (isValidCarId) {
+        let car = db.getById(carId);
+        showPromise(req, res, car);
+    } else {
+        res.status(400).json({error: 'invalid carId'});
+    }
 };
 
 exports.updateCarById = function(req, res) {
-    let carId = req.params.carId;
-    let carFields = helper.getCarFromReq(req);
-    let car = db.updateCarById(carId, car);
-    showPromise(req, res, car);
+    let carId = parseInt(req.params.carId, 10);
+    let isValidCarId = validation.validateCarId(carId);
+    if (isValidCarId) {
+        let carFields = helper.getCarFromReq(req);
+        let car = db.updateCarById(carId, car);
+        showPromise(req, res, car);
+    } else {
+        res.status(400).json({error: 'invalid carId'});
+    }
 };
 
 exports.deleteCarById = function(req, res) {
-    let result = db.deleteCarById(req.params.carId);
-    showPromise(req, res, result);
+    let carId = parseInt(req.params.carId, 10);
+    let isValidCarId = validation.validateCarId(carId);
+    if (isValidCarId) {
+        let result = db.deleteCarById(req.params.carId);
+        showPromise(req, res, result);
+    } else {
+        res.status(400).json({error: 'invalid carId'});
+    }
 };
 
 exports.getMostPopular = function(req, res) {
-    let amount = Number(req.params.amount);
+    let amount = parseInt(req.params.amount, 10);
     let cars = db.getMostPopular(amount);
     showPromise(req, res, cars);
 };
@@ -40,7 +58,7 @@ function showCar(req, res) {
 }
 
 function handleError(req, res) {
-    return error => res.status(500).json({error: error});
+    return error => res.status(400).json({error: error});
 }
 
 function showPromise(req, res, promise) {
